Rename: CookieService 위치 변경
